[#1] move away from pods because of https://github.com/ember-cli/ember-cli/issues/1634
