+++ addon/components/ember-tree.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from './template';
+import layout from '../templates/components/ember-tree';
 
 const {
   computed,
