+++ addon/components/ember-tree/if-helper.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from './template';
+import layout from '../../templates/components/ember-tree/if-helper';
 
 export default Ember.Component.extend({
   layout: layout
