+++ app/components/ember-tree.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-tree/components/ember-tree';
