--- app/components/ember-tree/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-cli-tree/components/ember-tree/component';
