+++ app/components/ember-tree/if-helper.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-tree/components/ember-tree/if-helper';
