+++ app/components/ember-tree/trigger-expand.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-tree/components/ember-tree/trigger-expand';
