--- app/components/ember-tree/trigger-expand/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-cli-tree/components/ember-tree/trigger-expand/component';
